+++ libs/examples/src/sunburst-chart/sunburst-chart-examples.module.ts
@@ -16,6 +16,7 @@
 
 import { NgModule } from '@angular/core';
 import { DtButtonModule } from '@dynatrace/barista-components/button';
+import { DtOverlayModule } from '@dynatrace/barista-components/overlay';
 import { DtSunburstChartModule } from '@dynatrace/barista-components/sunburst-chart';
 import { DtThemingModule } from '@dynatrace/barista-components/theming';
 import { DtExampleSunburstChartCustomColor } from './sunburst-chart-custom-color-example/sunburst-chart-custom-color-example';
@@ -29,6 +30,7 @@ import { DtFormattersModule } from '@dynatrace/barista-components/formatters';
     DtButtonModule,
     DtThemingModule,
     DtFormattersModule,
+    DtOverlayModule,
   ],
   declarations: [
     DtExampleSunburstChartDefault,